moving fake objects from the main impl file, adding schedualing properties where needed
