+++ shoebox/app/com/keepit/common/cache/FortyTwoCache.scala
@@ -6,6 +6,7 @@ import com.google.inject.{Inject, Singleton}
 import play.api.Play.current
 import scala.collection.mutable
 import play.api.libs.json._
+import play.api.Plugin
 import com.keepit.common.plugin.{SchedulingPlugin, SchedulingProperties}
 import com.keepit.common.healthcheck.{Healthcheck, HealthcheckError, HealthcheckPlugin}
 
@@ -28,7 +29,7 @@ class CacheStatistics {
 }
 
 // Abstraction around play2-memcached plugin
-trait FortyTwoCachePlugin extends SchedulingPlugin {
+trait FortyTwoCachePlugin extends Plugin {
   val stats: CacheStatistics
 
   private[cache] def onError(error: HealthcheckError) {}
