+++ shoebox/app/com/keepit/common/mail/MailToKeepPlugin.scala
@@ -222,9 +222,3 @@ class MailToKeepPluginImpl @Inject()(
     scheduleTask(actorFactory.system, 10 seconds, 1 minute, actor, FetchNewKeeps)
   }
 }
-
-class FakeMailToKeepPlugin extends MailToKeepPlugin with Logging {
-  def fetchNewKeeps() {
-    log.info("Fake fetching new keeps")
-  }
-}
